+++ src/webComponents/components.ts
@@ -1,7 +1,6 @@
 import { $id } from "../utils/ui/domQuery";
 import PrettyDialog from "./dialog/PrettyDialog";
 import FormField from "./form/FormField";
-import ImageUploader from "./form/ImageUploader";
 import PrettyButton from "./form/PrettyButton";
 import TextInput from "./form/TextInput";
 import GameControl from "./gameplay/GameControl";
@@ -35,7 +34,6 @@ const components: ComponentRegistration[] = [
     ],
   },
   { tag: "text-input", Constructor: TextInput },
-  { tag: "image-uploader", Constructor: ImageUploader },
   { tag: "form-field", Constructor: FormField },
   { tag: "social-link", Constructor: SocialLink },
   { tag: "text-link", Constructor: TextLink },